added a new Endpoint "/reviews/findByGameIdAndOrderByLikes/{gameId}/{pageNumber}" that can be used for getting most liked reviews for agame in page size 10

+++ src/main/java/com/KociApp/RateGame/review/ReviewRepository.java
@@ -16,6 +16,6 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
     Optional<Review> findByUserIdAndGameId(Long userId, int gameId);
 
     @Query("SELECT r FROM Review r JOIN ReviewLike rl ON r.id = rl.review.id WHERE r.game.id = :gameId AND rl.likeDislike = true GROUP BY r.id ORDER BY COUNT(rl.id) DESC")
-    List<Review> findTopLikedReviewsByGameId(Long gameId, Pageable pageable);
+    List<Review> findTopLikedReviewsByGameId(int gameId, Pageable pageable);
 
 }
